@@ -1,9 +1,7 @@
 cmake_minimum_required(VERSION 3.15)
 project(LightStripAudioSync LANGUAGES CXX)
 
-set(BROADCAST_ADDRESS "192.168.0.233")
-
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 set(SOURCES
