@@ -1,47 +1,65 @@
 #pragma once
 
+#define NOMINMAX
+
 #include <atomic>
-#include <stdint.h>
-#include <stdio.h>
+#include <mutex>
+#include <queue>
 #include <thread>
 #include <vector>
 #include <winsock2.h>
-#include <Ws2tcpip.h>
 
 // Link with ws2_32.lib
 #pragma comment(lib, "Ws2_32.lib")
 
 class DataSender {
     private:
-        constexpr static size_t magnitudes_size       = 2;
-        constexpr static unsigned short port          = 3333;
-        constexpr static const char* discover_message = "DISCOVER_LIGHTSTRIP_AUDIOSYNC_DEVICE";
-        constexpr static const char* register_message = "REGISTER_LIGHTSTRIP_AUDIOSYNC_DEVICE";
-
-    private:
-        WSADATA wsa_data                      = {};
-        SOCKET socket                         = {};
-        std::vector<sockaddr_in> destinations = {};
-
-        std::atomic<bool> discover_thread_running = false;
-        std::thread* discover_thread_instance     = nullptr;
-
-        std::atomic<bool> listen_thread_running = false;
-        std::thread* listen_thread_instance     = nullptr;
-
-        uint8_t last_magnitudes[magnitudes_size] = {};
-        unsigned identical_magnitudes_count      = 0;
+        constexpr static unsigned short port       = 3333;
+        constexpr static const char* data_discover = "DISCOVER_LIGHTSTRIP_AUDIOSYNC_DEVICE";
+        constexpr static const char* data_register = "REGISTER_LIGHTSTRIP_AUDIOSYNC_DEVICE";
 
     public:
-        DataSender(void);
+        struct QueuedData {
+            public:
+                std::shared_ptr<uint8_t[]> data = nullptr;
+                size_t data_size                = 0;
+
+                enum class destination_t : uint8_t {
+                    broadcast = 0,
+                    device
+                } destination = destination_t::broadcast;
+        };
+
+    private:
+        WSADATA wsa_data = {};
+        SOCKET socket    = INVALID_SOCKET;
+
+        std::atomic<bool> listen_thread_is_running          = false;
+        std::unique_ptr<std::thread> listen_thread_instance = nullptr;
+
+        std::atomic<bool> send_thread_is_running          = false;
+        std::atomic<bool> send_thread_has_queued_data     = false;
+        std::unique_ptr<std::thread> send_thread_instance = nullptr;
+
+        std::atomic<bool> discover_thread_is_running          = false;
+        std::unique_ptr<std::thread> discover_thread_instance = nullptr;
+
+        std::mutex thread_mutex;
+        std::queue<QueuedData> send_queue     = {};
+        std::vector<sockaddr_in> destinations = {};
+
+        bool send(const uint8_t* data, size_t data_size, sockaddr_in& address);
+
+    public:
+        DataSender(void) = default;
         ~DataSender(void);
 
         int initialize(void);
-        int initialize_device(PCSTR destination_ip);
+        int initialize_device(const char* destination_ip);
 
-        static void discover_thread(DataSender* data_sender, std::atomic<bool>* running);
-        static void listen_thread(DataSender* data_sender, std::atomic<bool>* running);
+        static void listen_thread(DataSender* data_sender);
+        static void send_thread(DataSender* data_sender);
+        static void discover_thread(DataSender* data_sender);
 
-        int send_discover(void);
-        int send_magnitudes(uint8_t left, uint8_t right);
+        void enqueue(const uint8_t* data, size_t data_size, const QueuedData::destination_t destination);
 };
